Renames emilia_sfc_data.py
